refactor: extract PromptBubbleBuilder and add isCLINotFoundError to classifier
- Create PromptBubbleBuilder with buildBubbleHtml() and escapeHtml()
  for pure HTML context-chip rendering without UI dependencies
- Move isCLINotFoundError() from ChatToolWindowContent to PromptErrorClassifier
  (exception cause-chain walking is classification logic)
- ChatToolWindowContent delegates to both, removing appendHtmlChar entirely
